+++ src/error.rs
@@ -6,8 +6,8 @@ pub type WFResult<T> = Result<T, WFError>;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub enum WFError {
-    CommandIO,
-    CommandParse,
+    CommandIO(String),
+    CommandParse(String),
     CommandErr(String),
     WifiAction(String),
     HotspotCreate(String),